using native type
git-svn-id: https://ncisvn.nci.nih.gov/svn/cagrid/trunk@11478 dd7b0c16-3c94-4492-92de-414b6a06f0b1

+++ cagrid-1-0/caGrid/projects/data/src/java/utilities/gov/nih/nci/cagrid/data/utilities/DomainModelUtils.java
@@ -150,7 +150,7 @@ public class DomainModelUtils {
                         // if we are only interested in primary concepts, need
                         // to check the order too
                         if (primaryConceptOnly) {
-                            if (sm.getOrder() == 0) {
+                            if (sm.getOrder().intValue() == 0) {
                                 result.add(c);
                                 break;
                             }
